requirements.txt
certifi==2018.8.24
chardet==3.0.4
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11
django-bootstrap3==11.0.0
django-bootstrap4==0.0.7
django-forms-bootstrap==3.1.0
django-heroku==0.3.1
django-likes==2.0.0
django-registration==2.4.1
django-secretballot==1.0.0
django-tinymce==2.7.0
gunicorn==19.9.0
idna==2.7
peewee==3.7.0
Pillow==5.3.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-dateutil==2.7.3
python-decouple==3.1
pytz==2018.5
pyuploadcare==2.5.0
requests==2.19.1
six==1.11.0
urllib3==1.23
whitenoise==4.1
